Refactor eval violations form into standalone component (out of page)
